fix: إصلاح خطأ النموذج غير المدعوم والإحداثيات العشرية
- تغيير النموذج الافتراضي إلى Qwen2.5-VL-7B-Instruct (مدعوم على HuggingFace Router)
- إعادة ترتيب قائمة النماذج لوضع النماذج المستقرة أولاً
- إصلاح AIAction: تغيير x وy من Int? إلى Double? لمنع NumberFormatException
  عند إرجاع النموذج إحداثيات عشرية مثل 0.500
- إضافة منطق تحويل الإحداثيات: إذا كانت القيمة بين 0.0 و1.0 تُضرب
  في أبعاد الشاشة للحصول على موضع البكسل الفعلي
- إصلاح منطق الـ fallback: عند فشل أي نموذج بسبب عدم الدعم يُجرَّب
  التالي في القائمة تلقائياً (بدلاً من العودة للافتراضي فقط)
- تحسين extractJson: دعم كتل markdown مثل ```json ... ```
- تحسين buildSystemPrompt: توضيح صريح بأن x وy يجب أن تكون
  أعداداً صحيحة تمثل بكسل وليس قيماً مُعيَّرة

+++ app/src/main/java/com/aicontrol/app/ai/models/AIAction.kt
@@ -4,8 +4,9 @@ import com.google.gson.annotations.SerializedName
 
 data class AIAction(
     @SerializedName("action") val action: String,
-    @SerializedName("x") val x: Int? = null,
-    @SerializedName("y") val y: Int? = null,
+    // Double? بدلاً من Int? لأن بعض النماذج ترجع إحداثيات عشرية أو مُعيّرة (0.0–1.0)
+    @SerializedName("x") val x: Double? = null,
+    @SerializedName("y") val y: Double? = null,
     @SerializedName("text") val text: String? = null,
     @SerializedName("direction") val direction: String? = null,
     @SerializedName("reason") val reason: String = "",
